updated login vaadin designer files
