--- frontend/src/views/login-view.js
@@ -1,34 +1,0 @@
-import {html, PolymerElement} from '@polymer/polymer/polymer-element.js';
-import './google-signin.js';
-import '@vaadin/vaadin-ordered-layout/src/vaadin-vertical-layout.js';
-import '@vaadin/vaadin-text-field/src/vaadin-text-area.js';
-
-class LoginView extends PolymerElement {
-
-    static get template() {
-        return html`
-<style include="shared-styles">
-                :host {
-                    display: block;
-                    height: 100%;
-                }
-            </style>
-<vaadin-vertical-layout style="width: 100%; height: 100%; align-items: center;" id="vaadinVerticalLayout">
-  Welcome to Scrumptious 
- <google-signin id="signin" style="align-self: center;"></google-signin>
-</vaadin-vertical-layout>
-`;
-    }
-
-    static get is() {
-        return 'login-view';
-    }
-
-    static get properties() {
-        return {
-            // Declare your properties here.
-        };
-    }
-}
-
-customElements.define(LoginView.is, LoginView);
